requirements.txt
faster-whisper==1.2.1
flask==3.1.2
flask-cors==6.0.1
opencc-python-reimplemented==0.1.7
